Refactor scanner module and introduce new dispatch system
- Restructured the `scanner` module by splitting functionality into `full` and `changed` submodules for better organization and clarity.
- Introduced a new `dispatch` module to handle command execution for various CLI commands, improving command management.
- Updated imports across multiple files to reflect the new module structure and removed unused imports to clean up the codebase.
- Added new functionality for computing blast radius and content signals in the review process, enhancing the analysis capabilities.
- Implemented CI report generation for review reports, allowing for better integration with continuous integration workflows.
- Enhanced the rendering of scan summaries and baseline reports with new output formats, improving user experience.
